requirements.txt
accelerate==0.21.0
datasets==2.14.0
huggingface-hub==0.16.4
pytest==7.4.0
python-dotenv==1.0.0
tokenizers==0.13.3
torch==2.0.1
transformers==4.31.0
typed-argument-parser==1.8.1